content/html/content/test/test_bug546995-3.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=546995
-->
<head>
  <title>Test for Bug 546995</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<!-- When the page is loaded, we wait 2 seconds before launching the tests
     because the focus events may not be finished when the page is loaded so
     we have to wait or launch the tests when a focus events is thrown.
     However, if autofocus is ignored, there will be no focus event so we can't
     rely on that. -->
<body onload="window.setTimeout(runTests, 2000);">
<script type="application/javascript">
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  var prefs = Components.classes["@mozilla.org/preferences-service;1"]
            .getService(Components.interfaces.nsIPrefBranch);
  var gAutofocusPref = prefs.getBoolPref("browser.autofocus");
  prefs.setBoolPref("browser.autofocus", false);

  var gFocusHandled = false;
</script>
</script>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=546995">Mozilla Bug 546995</a>
<p id="display"></p>
<div id="content">
  <input autofocus onfocus='gFocusHandled=true;'>
  <textarea autofocus onfocus='gFocusHandled=true;'></textarea>
  <select autofocus onfocus='gFocusHandled=true;'></select>
  <button autofocus onfocus='gFocusHandled=true;'></button>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 546995 **/

// TODO: keygen should be added when correctly implemented, see bug 101019.

SimpleTest.waitForExplicitFinish();

function runTests()
{
  // We do not check |document.activeElement| to be sure the input is never
  // ever focused.
  ok(!gFocusHandled,
     "When browser.autofocus=false, autofocus attribute should be ignored");

  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  prefs.setBoolPref("browser.autofocus", gAutofocusPref);

  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
